Break out specs into seperate files
